# schulapp-backend/openapi.yaml
openapi: "3.1.0"
info:
  title: Schul-App API
  version: "0.1.0"
  description: |
    API-Spezifikation der Schul-App.
    Diese Datei ist die einzige Quelle der Wahrheit für alle Endpunkte.
    `make generate` erzeugt daraus Go-Handler-Interfaces und TypeScript-Client.
servers:
  - url: /api/v1

paths:
  /health:
    get:
      operationId: healthCheck
      summary: Health check
      responses:
        "200":
          description: Server is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/login:
    post:
      operationId: login
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token via HttpOnly refresh-token cookie
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                required: [access_token]
                properties:
                  access_token:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/logout:
    post:
      operationId: logout
      summary: Logout and invalidate refresh token
      responses:
        "204":
          description: Logged out successfully

  /auth/me:
    get:
      operationId: getMe
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateMe
      summary: Update current user profile (name only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1

    LoginResponse:
      type: object
      required: [access_token, user]
      properties:
        access_token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    UpdateMeRequest:
      type: object
      required: [first_name, last_name]
      properties:
        first_name:
          type: string
          minLength: 1
        last_name:
          type: string
          minLength: 1

    User:
      type: object
      required: [id, email, first_name, last_name, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]